move list search (component) to its own file
